detector/requirements.txt
flask
pandas
requests
sqlalchemy
joblib
scikit-learn
psycopg2-binary